Api updates

+++ objathor/utils/download_utils.py
@@ -5,7 +5,11 @@ from filelock import FileLock
 from tqdm import tqdm
 
 
-def download_with_progress_bar(url: str, save_path: str, desc: str = ""):
+def download_with_progress_bar(url: str, save_path: str, desc: str = "", overwrite: bool = False):
+    if os.path.exists(save_path) and not overwrite:
+        print(f"File already exists: {save_path}")
+        return
+
     with open(save_path, "wb") as f:
         print(f"Downloading {save_path}")
         response = requests.get(url, stream=True)
